perf(dashboard): trim hot path imports
